preparation of G4 user actions for adept GPU project
